split ReplyController into Sender and Receiver for easier reasoning
